requirements.txt
batchgenerators==0.25
certifi==2021.10.8
charset-normalizer==2.1.0
click==8.1.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama==0.4.6
commonmark==0.9.1
cycler==0.11.0
cytoolz==0.11.0
dask==2021.10.0
Deprecated==1.2.13
dicom==0.9.9.post1
docopt==0.6.2
dominate==2.7.0
et-xmlfile==1.1.0
filelock==3.9.0
fonttools==4.33.3
fsspec @ file:///opt/conda/conda-bld/fsspec_1647268051896/work
future==0.18.3
huggingface-hub==0.12.1
humanize==4.4.0
idna==3.3
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
importlib-metadata==6.0.0
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.2
linecache2==1.0.0
locket @ file:///tmp/build/80754af9/locket_1646988443534/work
matplotlib==3.5.2
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monai==1.1.0
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
nibabel==3.2.2
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1649782762302/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.5.64
openpyxl==3.1.2
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
Pillow==9.5.0
pipreqs==0.4.13
protobuf==4.24.4
pydicom==2.3.0
Pygments==2.14.0
pynrrd==0.4.3
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil==2.8.2
pytorch-metric-learning==2.1.1
pytz==2022.2.1
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1648727874099/work
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
rich==12.6.0
scikit-image @ file:///tmp/build/80754af9/scikit-image_1648196043170/work
scikit-learn==1.0.2
scipy @ file:///tmp/build/80754af9/scipy_1618852618548/work
seaborn==0.12.0
shellingham==1.5.0.post1
SimpleITK==2.1.1.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2020.10.1
timm==0.6.12
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.10.0
torchaudio==0.10.0
torchfile==0.1.0
torchio==0.18.86
torchvision==0.11.0
tornado==6.2
tqdm==4.64.0
traceback2==1.4.0
typer==0.7.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1650247814259/work
unittest2==1.1.0
urllib3==1.26.11
visdom==0.1.8.9
websocket-client==1.3.3
wrapt==1.14.1
yarg==0.1.9
zipp==3.11.0
